I did find the bug(s) with attributed variables.
What apparently needs to be done on unifying to an attribute variable is

1) Build interrupts.  For an attributed variable, this means passing a
pointer to the attribute list.

2) Bind the attributed variable "eagerly" (as opposed to lazily, as
before)

This was being done correctly for the Prolog unification macros, but
not for the trie unification macros.  These latter macros passed a
pointer to the attributed variable (not the attr list) and skipped
step 2.

I changed what seemed to be appropriate places in tr_code_xsb_I.h,
namely in unify_with_trie_numcon, unify_with_trie_str,
unify_with_trie_list and unify_with_trie_attv.  However I did not
change unify_with_trie_val, as the comments in the code (from Bao, I
think) indicated that no interrupt should be dispaltched in this case.

This seemed to fix things.  The problem observed in a previous email
was that when the trie instructions passed a pointer to an attributed
variable, rather than to an attribute list, unification insructions in
the prolog attribute handling caused new bindings (e.g. binding to [])
causing new interrupts and an infinite loop.