Refactored init_machine() to include a new function init_thread_structures() that includes all initializations that need to be done on a per-thread basis.  Then mem_dealloc'ed all thread-specific memory, including varstrings and such in cleanup_tread_structures() (nee cleanup_machine()).  This change should affect the MT engine only.
At this point, there should be no memory leaks simply for
creating a thread that exits -- apart from interaction with the
string table, etc.  Also, I'm not freeing space for dispatch
blocks for private predicates, which doesn't seem necessary
overall.  I suppose if we want to get rid of that space, the
reclamation should be done in one of the abolish predicates,
rather than xsb_thread_exit()?