debug_xsb.c -- changes to allow print_delay_list from generic debug config emuloop.c -- removed call to gdb_dummy emudef.h --
Change to  affect the ordering of checks for interrupt vectors
at call and proceed.  We had been checking for spy points before
checking for attributed variable interrupts, and I reversed the order,
allowing me to use the debugger/tracer on CHR programs (still not a
great way to debug, but better than nothing for now).  I've been using
this change for a couple of days, and it seems to have no effect (as
indeed it shouldn't) but please let me know if it seems to be changing
anything in your code.