Modified slide_heap() and copy_heap() by adding a new function trail_hp_pointer_from_cell() used in the chaining portion of slide_heap().
hp_pointer_from_cell() returns the address,tag pair if a cell points
to part of the heap.  The trail needs a special function for this
because of pre-image trailing used for handling attributed variables.
The pre-image mark is handled correctly by the garbage collectors --
information about this is set in the array gc_marks in the marking
phase so that subsequent phases deal with clean addresses.  However,
the pre-image value may be a list (or structure) that is in the
attv-interrupt vector rather than in the heap itself.  The new
function thus double-checks to see whether structures or lists are
actually in the heap.

I made a similar change of hp_pointer_from_cell() to
trail_hp_pointer_from_cell() in copy_heap() in gc_copy.h

I also added a new test to the test suite to test this situation.