These changes implement a facility that allows a routine called from XSB as foreign code to call a predicate back in XSB.  This allows mutual recursion between XSB and foreign code (and has been tested by programming a recursive fibonacci in which the Prolog code calls C for the samller values and C calls Prolog for its smaller values.)
It is similar to the "C calling XSB" interface, with the following
differences:

1. Do not call xsb_init or xsb_init_string.  These functions
initialize XSB, which clearly cannot be done if XSB is already
running.

2. A foreign routine called from XSB that wants to call an XSB
predicate must first call:
	xsb_query_save(NumRegs)

It must do this after retrieving its input parameters from XSB and
before calling any XSB predicate.  NumRegs must be the number of XSB
registers that the current foreign code uses (i.e., the number of
arguments of the XSB predicate used to call this foreign code
routine.)  This call saves the caller registers and sets up the XSB
state to accept queries from the foreign code.  (This is similar to
the XSB state after xsb_init is called.)  So after this call, the
foreign routine can use any of the interface routines for calling XSB,
such as xsb_command, xsb_command_string, xsb_query, xsb_query_string,
etc.

3. Before setting its output parameters and returning to its caller,
the foreign routine must call
	xsb_query_restore()

This restores the XSB parameter registers back to their state when
xsb_query_save() was called, and so prepares the registers to accept
the return values.  (It must be called even if there are no values to
be returned.)

This code works (according to my tests) for the single-threaded
engine.  It does NOT currently work for the multi-threaded engine.
I've put in the necessary CTXT- stuff, but still get crashes.  I'm
sure it can be made to work, but may need help in figuring out how.