Refactored assert code to make trie-assert faster.  The differences vary, from about 15% on my old mac to 200+% on my new mac (!).  In doing so I did not slow down assert code except just a little to fix a minor bug -- it turned out that trie assert wasn't properly checking that clauses with bodies were not being asserted.  This meant that if p(X) were trie indexed, then p(1) would have trie indexing, but p(X):- f(X) would not.
The assert code is slowed down only minimally, but there is much
that can be done to speed up assert significantly (and
trie_assert more), like moving convert_to_dyna and check_assert
into C, both of which should be easy.

Also, a very minor efficiency factoring of retract.  Just as with
assert, much can be done to make retract more efficient, though
this is less important, as our retractall is very fast.