Fixed recently introduced bug in mutex_test.  While I was in mutex_xsb, I decided to fix an outstanding bug where we could get core dumps if a mutex is called with an integer that is not a pointer to a dynamic mutex structure.  I fixed this by making sure there is an alias record for each mutex -- this record has the pointer in the first argument as long as it is valid.
The result for this is that access to user mutexes is somewhat slower
than before for integer pointers.  It had taken about 0.7 secs for
1000000 locks and unlocks, while now we're at about 1.1 secs.  Access
through aliases, however, is faster, again 1.1 secs down from 1.6
secs.

Ideally, we wouldn't use Prolog for aliases, which would allow alias
and other error checking to be handled in C.  This would speed up
mutexes, user-level tries, and to some extent message queues and
stream operations.  But I don't have time for this right now.

Terry