Two small changes.
First, I made sure that the catch in the command-line interpreter
resets the input stream.  The problem is that if read, fmt_read,
etc throw an exception, XSB exits which is annoying.  I put the
see/1 in the catch itself rather than the default error handler
since I only want to reset the stram when the interpreter catches
it.

Second, I took gc out of the banner.  The banner was starting to
look a little verbose, and 99.99% of the time we call indirect gc
anyway.