A user rightly complained about error messages in univ.  I restructured univ code a little so that univ and our error reporting for it is ISO-compliant for non-Hilog predicates (well, almost).
The changes for ISO compliance mean that univ will now throw an
error in a very few cases where it either used to fail or
succeed (e.g. X =.. [3|Y] used to succeed and now throws an
error).

Let me know if the changes turn out to be a problem for the
existing XSB code base.

Terry