Changes were mostly for documentation and code refactoring.
However, also fixed a latent bug in simplification where
was_simplifiable was not properly checking negative delay elements
corresponding to subsumed subgoal frames.  This led to the possibility
of adding un-needed delay elements that might not have been simplified
away.  Surprisingly enough, however, none of the tests in the test
suite checked for this.